Fixed bug at time management; Include pthread header for compatibility with mingw; Fixed bug at pv collector.
git-svn-id: https://svn.code.sf.net/p/redqueenchess/code/redqueen@116 865e88f3-cc1b-4e5f-86d1-2a7dcd23a5c5

+++ src/SearchAgent.cpp
@@ -105,7 +105,7 @@ void* SearchAgent::startThreadSearch() {
 		simplePV.setSearchFixedDepth(true);
 		simplePV.setDepth(defaultDepth);
 	}
-	// TODO implement movestogo and nodes
+
 	simplePV.search();
 	return 0;
 }
@@ -142,10 +142,10 @@ const uint32_t SearchAgent::getTimeToSearch() {
 		return this->getMoveTime();
 	}
 
-	uint32_t time=board.getSideToMove()==WHITE ?  this->getWhiteTime() : this->getBlackTime();
+	int time=board.getSideToMove()==WHITE ?  this->getWhiteTime() : this->getBlackTime();
 	//uint32_t incTime=board.getSideToMove()==WHITE ? this->getWhiteIncrement() : this->getBlackIncrement();
 
-	size_t movesLeft = movesToGo>0?movesToGo : defaultGameSize-board.getMoveCounter();
+	int movesLeft = movesToGo>0?movesToGo : defaultGameSize-board.getMoveCounter();
 
 	if (movesLeft<=1) {
 		movesLeft=defaultGameSizeInc;
